Initial upload
This is what I've been working on
All of the routine functions work to my knowledge
I need the tokenizer to be able to read non-parenthetical functions
before I start testing recursive cases.

So far it works pretty well, I really recommend testing the class
manually and tracing the execution with breakpoints.